requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
gTTS==2.2.3
idna==3.3
multidict==5.2.0
pycodestyle==2.8.0
pycparser==2.21
PyNaCl==1.4.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQuendo==1.0.0
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
youtube-dl==2021.12.17
